settings cache, fade to black and gles2
implemented a fade to black system using the new layer color api that is used on scene switch.
This mitigates the problem of intial shader compile (of the immediatly visible objects) as
this now happens when the screen is black.

Since GLES2 performs at the moment significantly better then gles3 it is now the default. The
implication of this is that now ETC needs to be used which does not support alpha channels so
the splash screen now has a black background instead of transparent.

Settings set using the vr.set_XXX helper functions will now be cached and restored on app resume.